refactor(Compiler): Rewrite of module resolving and compiling.
Still needs some performance improvements, feels kinda slow.

The goal of this rewrite was to make a more defined line between something that can be resolved, something that is completed and has been compiled and ensuring methods aren't too long and are kept to their respective classes.

This moves alot of logic around and will definite have some stuff it has broken but should be easier to work with in the future.