feat(Logging): Enhance error formatting with console width handling and ellipsis.
Ensures that the error fits on a single line and each side is trimmed to an even amount to supply the most context around the statement.

Techinally if the statement is large enough it will start chopping it off at the end, but thats kind of an edge case i dont want to deal with for now.